requirements.txt
fastapi
python-fasthtml
firebase-admin
python-multipart
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
